create_collection prework
