+++ src/db/action.rs
@@ -1,0 +1,1 @@
+pub(crate) mod create_collection;
